hayes-protocol/backend/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
python-dotenv==1.2.2
pydantic==2.12.5
google-genai>=1.0.0
openai>=1.0.0
